requirements.txt
streamlit==1.37.0
faiss-cpu==1.8.0
sentence-transformers==3.0.1
transformers==4.43.3
accelerate==0.33.0
bitsandbytes==0.43.1      # optional, speeds some models; ignore if it errors
pypdf==4.2.0
python-dotenv==1.0.1
openai==1.40.0
langdetect==1.0.9


